+++ generators/app/templates/node/lib/app-lib.js
@@ -1,11 +1,17 @@
 "use strict";
 
-function test(arg)
+function test1(arg1, arg2)
 {
-    return "Arg is: " + arg;
+    return "1: Arg is: " + arg;
+}
+
+function test2(arg2_1, arg2_2)
+{
+    return "2: Arg is: " + arg;
 }
 
 module.exports =
 {
-    test: test
+    test1: test1,
+    test2: test2
 };